AI now threaded
Does that mean it's screwed? :stuck_out_tongue:

Damn that was so bad it hurt to type

Regardless, AI is threaded now and will constantly run in the background until the program ends. Will consider adding a function to stop all threads to prevent overtraining but if we keep calling Train() on new data it should ideally prevent the issue (or run us out of memory, one of the two). TODO: Make it so you can save weights, working on that now